refactor : motionevent 수정
