+++ app/src/main/java/View/DrawObjects.kt
@@ -1,10 +1,11 @@
 package View
 
-import Data.Plane
+import Presenter.Plane
 import Data.Rectangle
 import android.content.Context
 import android.graphics.Canvas
 import android.util.AttributeSet
+import android.util.Log
 import android.view.MotionEvent
 import android.view.View
 
@@ -32,17 +33,29 @@ class DrawObjects(context: Context?, attrs: AttributeSet?) : View(context, attrs
     override fun onTouchEvent(event: MotionEvent?): Boolean {
         val x = event?.x
         val y = event?.y
+        var foundRectangle: Boolean = false
         when (event?.action) {
             MotionEvent.ACTION_DOWN -> {
                 rectangles.forEach { rectangle ->
-                    var foundRectangle = plane.isOnRectangle(x, y, rectangle)
-                    if(foundRectangle) {
+                    foundRectangle = plane.isOnRectangle(x, y, rectangle)
+                    if (foundRectangle) {
+                        Log.d("rectangle right", "rectangle on the pic")
+                        if(plane.rectangleStrokeList.contains(rectangle)) {
+                            return@forEach
+                        }
                         rectangleStrokes = plane.getRectangleStrokesList(foundRectangle, rectangle)
+                        return@forEach
+                    } else {
+                        plane.setRectangleStrokeListEmpty()
+                        rectangleStrokes.clear()
                     }
                 }
                 invalidate()
             }
+//            MotionEvent.ACTION_UP -> {
+//              TODO("커스텀리스너구현하기")
+//            }
         }
-        return super.onTouchEvent(event)
+        return true
     }
 }
