+++ app/src/main/java/View/MainActivity.kt
@@ -1,16 +1,9 @@
 package View
 
-import Data.Plane
-import Data.Rectangle
-import android.content.Context
-import android.graphics.*
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.AttributeSet
-import android.view.MotionEvent
 import android.widget.Button
 import com.example.app.R
-import android.view.View as View
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
